requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
circuitbreaker==2.1.3
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==5.0.1
hyx==0.0.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-consul==1.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tenacity==9.1.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3